requirements.txt
tcod>=13.8.1
numpy>=1.24.0
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
